feat(plot): decompose plot page orchestrator
