+++ docs/ai-logs/2026-05-27-issue-627.md
@@ -1,0 +1,26 @@
+# 2026-05-27 — Refactor page orchestrators (Issue 627)
+
+## Session Narrative
+
+### Prompt 1: Decompose Plot Page Orchestrator
+
+**AI response**:
+- Created `src/lib/state/plot-page-orchestrator.svelte.ts` to manage the plot page state
+- Lifted state initialization, mode selection logic, and WASM readiness reactions from `plot/+page.svelte` into the new orchestrator
+- Refactored `src/routes/plot/+page.svelte` template to utilize the new orchestrator methods
+- Updated the page-init contract tests in `src/tests/contracts/page-init.contract.test.ts` to test the new plot orchestrator
+- Completed `task.md` checkboxes
+- Formatted, checked, and tested the codebase ensuring 100% pass rate
+
+## Tasks
+
+### Plot Page Decomposition
+
+- **Status**: completed
+- **Stage**: Technical Debt / Maintenance
+- **Files changed**:
+  - `src/lib/state/plot-page-orchestrator.svelte.ts`
+  - `src/routes/plot/+page.svelte`
+  - `src/tests/contracts/page-init.contract.test.ts`
+- **Decision**: Methods used in callbacks were rewritten as arrow functions inside the template to ensure `this` stays bound to the `PlotPageOrchestrator` instance.
+- **Issue**: None
